Redid event system to pipe through the event manager. Now there are publishers, event types, and subscibers.
Publishers publish to the manager, filing under an event type.
Subscribers lists to the manager for specific event types.

However there are convenience methods on the publishers in order to hide the fact that the manager is orchestrating everything.

Additionally, support for delayed events was added.
For instance, if events need to be processed at the end of the current update logic, that can be implemented.

Somehow someway, a mash of inheritence and generics allows for near perfect type safety in dealing with the event system. :D / D:

As such, copious documentation should be put forth in managing the event system.

[It's not the worst code I have ever seen though.]
[In fact I would trade simple code for type safety any day]